HomeWork06 - ATM Emulator

+++ HW06_ATM_EMULATOR/src/main/java/ru/otus/hw06/atm/service/BankATMService.java
@@ -4,14 +4,13 @@ import ru.otus.hw06.atm.model.InsufficientFundsException;
 import ru.otus.hw06.atm.model.Nominal;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Map;
 
 public interface BankATMService extends Serializable {
 
     void addMoney(Map<Nominal, Long> banknotes);
 
-    Map<Nominal, Long> getMoney(BigDecimal moneySum) throws InsufficientFundsException;
+    Map<Nominal, Long> getMoney(long moneySum) throws InsufficientFundsException;
 
-    BigDecimal getCurrentMoneySum();
+    long getCurrentMoneySum();
 }
